// src/app/models/residence/residence.ts
export class Residence {
  date: string;
  firstName: string;
  lastName: string;
  fatherName: string;
  nation: string;
  passportNo: string;
  address: string;
  phone: string;
  introduced: string;
  image: string;
  passportImage: string;
  requestForm: string;
  familyFirstName: string;
  familyLastName: string;
  familyNation: string;
  familyPhone: string;
}